appengine/predator/app/requirements.txt
appengine-python-standard==1.1.1
Flask==1.0.2
google-api-python-client==2.2.0
google-cloud-storage==1.44.0
gunicorn==20.1.0
httplib2==0.19.1
numpy==1.21.1
oauth2client==4.1.3

# Needed by the above
attrs==21.4.0
cachetools==3.1.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
crcmod==1.7
frozendict==2.0.6
google-api-core==1.31.2
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-crc32c==1.5.0
google-resumable-media==1.3.3
googleapis-common-protos==1.52.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mock==4.0.3
packaging==20.9
Pillow==8.3.1
protobuf==3.20.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==2.4.7
pytz==2023.3.post1
requests==2.27.1
rsa==4.5
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
setuptools==44.1.1
six==1.16.0
urllib3==1.26.16
Werkzeug==1.0.1
